+++ spec/model_spec.rb
@@ -102,6 +102,74 @@ RSpec.describe AttrJson::Record do
     end
   end
 
+  describe "#dup" do
+    let(:nested_class) do
+      Class.new do
+        include AttrJson::Model
+
+        attr_json :str, :string
+      end
+    end
+    let(:klass) do
+      nested_class_type = nested_class.to_type
+      Class.new do
+        include AttrJson::Model
+
+        attr_json :str_array, :string, array: true
+        attr_json :str, :string
+        attr_json :nested, nested_class_type, default: {}
+      end
+    end
+
+    it "dups all nested data" do
+      obj = klass.new(str: "one", str_array: ["a", "b"], nested: nested_class.new(str: "two"))
+      obj_dup = obj.dup
+
+      expect(obj_dup.str).not_to equal(obj.str)
+      expect(obj_dup.str_array).not_to equal(obj.str_array)
+      expect(obj_dup.str_array.first).not_to equal(obj.str_array.first)
+
+      expect(obj_dup.nested).not_to equal(obj.nested)
+      expect(obj_dup.nested.str).not_to equal(obj.nested.str)
+    end
+  end
+
+  describe "#freeze" do
+    let(:nested_class) do
+      Class.new do
+        include AttrJson::Model
+
+        attr_json :str, :string
+      end
+    end
+    let(:klass) do
+      nested_class_type = nested_class.to_type
+      Class.new do
+        include AttrJson::Model
+
+        attr_json :str_array, :string, array: true
+        attr_json :str, :string
+        attr_json :nested, nested_class_type, default: {}
+      end
+    end
+
+    it "makes it not possible to set attributes" do
+      obj = klass.new(str: "one", str_array: ["a", "b"], nested: nested_class.new(str: "two"))
+
+      obj.freeze
+
+      # ruby previous to 2.5 didn't have FrozenError
+      expected_error = defined?(FrozenError) ? FrozenError : RuntimeError
+
+      expect {
+        obj.str = "foo"
+      }.to raise_error(expected_error)
+
+      # note it's not actually deep-frozen, you can mutate attributes including
+      # arrays. deep freeze could be a different feature.
+    end
+  end
+
   describe "unknown keys" do
     let(:klass) do
       Class.new do
@@ -234,8 +302,23 @@ RSpec.describe AttrJson::Record do
       end
     end
 
-    it "available" do
+    it "available in registry" do
       expect(klass.attr_json_registry.attribute_names).to match([:str_one, :int_one])
+    end
+
+    it "available via .attribute_types" do
+      expect(klass.attribute_types).to eq({
+        "str_one" => ActiveModel::Type.lookup(:string),
+        "int_one" => ActiveModel::Type.lookup(:integer),
+      })
+    end
+
+    it "available via .attribute_names" do
+      expect(klass.attribute_names).to match([:str_one, :int_one])
+    end
+
+    it "available via #attribute_names" do
+      expect(klass.new.attribute_names).to match([:str_one, :int_one])
     end
   end